+++ lib/biquad_shelf.h
@@ -3,11 +3,11 @@
 
 class BiquadShelf : public Biquad {
 public:
-    BiquadShelf(float gain_db, int fc, int fs) {
+    BiquadShelf(float gain_db, float fc, int fs) {
         construct(gain_db, fc, fs);
     }
 
-    virtual tp_coeffs& calculate_coeffs(float gain_db, int fc, int fs) = 0;
+    virtual tp_coeffs& calculate_coeffs(float gain_db, float fc, int fs) = 0;
 
     void set_gain(float gain_db) {
         m_gain = gain_db;
@@ -39,7 +39,7 @@ protected:
     }
 
 private:
-    void construct(float gain_db, int fc, int fs) {
+    void construct(float gain_db, float fc, int fs) {
         m_gain = gain_db;
         m_fc = fc;
         m_fs = fs;
