+++ lib/fo_shelving_low.h
@@ -15,7 +15,7 @@ class FO_SHELVING_LOW : public BiquadShelf {
     using BiquadShelf::BiquadShelf;
 
 public:
-	tp_coeffs& calculate_coeffs(float gain_db, int fc, int fs)
+	tp_coeffs& calculate_coeffs(float gain_db, float fc, int fs)
 	{
 		coef_size_t th = 2.0 * pi * fc / fs;
 		coef_size_t m = pow(10.0, gain_db / 20.0);
