+++ lib/so_parametric_ncq.h
@@ -24,7 +24,7 @@ class SO_PARAMETRIC_NCQ : public BiquadParametric {
     using BiquadParametric::BiquadParametric;
 
 public:
-    tp_coeffs& calculate_coeffs(float gain_db, float Q, int fc, int fs)
+    tp_coeffs& calculate_coeffs(float gain_db, float Q, float fc, int fs)
     {
         coef_size_t w = 2.0 * pi * fc / fs;
         coef_size_t m = pow(10.0, gain_db / 20.0);
